Switched to TeX Documentation (#15)
Packaging for CTAN requires documentation to be in a specific PDF format, so I figured rather than duplicating between the markdown README and a new doc I'd just keep it centralised in one. This creates those new documents, updates the previous docs to reflect that, and adds scripts to compile them automatically.